Attempt to remove circular import
