+++ esp/esp/program/controllers/autoscheduler/db_interface.py
@@ -13,7 +13,7 @@ from django.db import transaction
 from esp.resources.models import \
     ResourceType, Resource, ResourceAssignment, ResourceRequest
 from esp.program.models import ClassSection
-from esp.program.models.class_ import ClassStatus
+from esp.program.models.classstatus import ClassStatus
 from esp.users.models import ESPUser, UserAvailability
 from esp.cal.models import Event
 from esp.program.modules import module_ext
