+++ esp/esp/program/models/classstatus.py
@@ -1,0 +1,6 @@
+class ClassStatus():
+    CANCELLED = -20
+    REJECTED = -10
+    UNREVIEWED = 0
+    HIDDEN = 5
+    ACCEPTED = 10
