+++ esp/esp/program/modules/handlers/onsiteclasslist.py
@@ -45,7 +45,7 @@ from django.utils.safestring import mark_safe
 
 from esp.users.models    import ESPUser, Record, ContactInfo, StudentInfo, K12School
 from esp.program.models import RegistrationProfile
-from esp.program.models.class_ import ClassStatus
+from esp.program.models.classstatus import ClassStatus
 
 from esp.program.modules.base import ProgramModuleObj, needs_onsite, needs_student, main_call, aux_call
 from esp.program.models import ClassSubject, ClassSection, StudentRegistration, ScheduleMap, Program
