+++ esp/esp/program/modules/handlers/programprintables.py
@@ -37,7 +37,7 @@ from esp.utils.web import render_to_response
 from esp.users.models    import ESPUser, User, Record, RecordType
 from esp.program.models  import ClassSubject, ClassSection, StudentRegistration
 from esp.program.models  import ClassFlagType
-from esp.program.models.class_ import ClassStatus
+from esp.program.models.classstatus import ClassStatus
 from esp.users.views     import search_for_user
 from esp.users.controllers.usersearch import UserSearchController
 from esp.utils.latex  import render_to_latex
