+++ esp/esp/program/modules/handlers/teachercheckinmodule.py
@@ -38,7 +38,8 @@ from esp.program.modules.base import ProgramModuleObj, needs_teacher, needs_stud
 from esp.program.modules import module_ext
 from esp.program.modules.handlers.grouptextmodule import GroupTextModule
 from esp.program.models import RegistrationProfile
-from esp.program.models.class_ import ClassSubject, ClassSection, ClassStatus
+from esp.program.models.class_ import ClassSubject, ClassSection
+from esp.program.models.classstatus import ClassStatus
 from esp.program.models.flags import ClassFlagType
 from esp.utils.web import render_to_response
 from esp.utils.decorators import json_response
