+++ esp/esp/users/models/__init__.py
@@ -78,7 +78,7 @@ from esp.utils.decorators import enable_with_setting
 from esp.utils.expirable_model import ExpirableModel
 from esp.utils.widgets import NullRadioSelect, NullCheckboxSelect
 from esp.utils.query_utils import nest_Q
-from esp.program.models.class_ import ClassStatus
+from esp.program.models.classstatus import ClassStatus
 
 from urllib import quote
